# openapi-spec.yaml
openapi: 3.0.3
info:
  title: 智慧乡村综合服务平台 API
  description: |
    智慧乡村综合服务平台的完整 API 规范，涵盖村民管理、村务治理、信息公示、生活服务四大核心模块。

    ## 架构说明
    - **主API服务器** (端口3001): 监控、稳定性管理、多语言支持、通知服务
    - **村务服务器** (端口5000): 核心村务管理、Socket.IO实时通信、紧急广播

    ## 核心功能模块
    1. **村委管理**: 村委信息管理、村务协同、任务调度
    2. **村民管理**: 信息数字化、在线办事、档案管理
    3. **村务治理**: 财务管理、项目管理、公开透明
    4. **信息公示**: 政策公告、村务通知、关键词查询
    5. **安全防护**: 数据加密、权限管控、操作审计
  version: 2.0.0
  contact:
    name: 智慧乡村平台团队
    email: support@smartvillage.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: 主API服务器 (监控、通知、多语言)
  - url: http://localhost:5000
    description: 村务服务器 (核心服务、Socket.IO)
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - 认证管理
      summary: 用户登录
      description: 支持多种登录方式：账号密码、手机验证码、人脸识别
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                loginType:
                  type: string
                  enum: [password, sms, face]
                  description: 登录方式
                username:
                  type: string
                  description: 用户名或手机号
                password:
                  type: string
                  description: 密码（password方式必需）
                smsCode:
                  type: string
                  description: 短信验证码（sms方式必需）
                faceData:
                  type: string
                  description: 人脸识别数据（face方式必需）
              required:
                - loginType
                - username
            examples:
              password_login:
                summary: 密码登录
                value:
                  loginType: "password"
                  username: "admin"
                  password: "password123"
              sms_login:
                summary: 短信验证码登录
                value:
                  loginType: "sms"
                  username: "13800138000"
                  smsCode: "123456"
              face_login:
                summary: 人脸识别登录
                value:
                  loginType: "face"
                  username: "13800138000"
                  faceData: "base64_encoded_face_data"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/register:
    post:
      tags:
        - 认证管理
      summary: 用户注册
      description: 新用户注册，需要手机验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  pattern: '^1[3-9]\d{9}$'
                  description: 手机号码
                name:
                  type: string
                  minLength: 2
                  maxLength: 20
                  description: 真实姓名
                password:
                  type: string
                  minLength: 6
                  description: 密码
                smsCode:
                  type: string
                  description: 短信验证码
                idCard:
                  type: string
                  pattern: '^[1-9]\d{5}(18|19|20)\d{2}((0[1-9])|(1[0-2]))(([0-2][1-9])|10|20|30|31)\d{3}[0-9Xx]$'
                  description: 身份证号码
                villageId:
                  type: string
                  description: 所属村庄ID
              required:
                - phone
                - name
                - password
                - smsCode
                - idCard
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  # ==================== 村委管理 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          description: JWT访问令牌
        refreshToken:
          type: string
          description: 刷新令牌
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            name:
              type: string
            role:
              type: string
            village:
              type: string
            permissions:
              type: array
              items:
                type: string
        expiresIn:
          type: integer
          description: 令牌有效期（秒）
      required:
        - success
        - token
        - user
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            name:
              type: string
            phone:
              type: string
            villageId:
              type: string
        message:
          type: string

    # ==================== 村委管理 ====================
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "请求参数错误"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "未授权访问"
              code:
                type: string
                example: "UNAUTHORIZED"
...
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "资源已存在"
              conflictField:
                type: string
    TooManyRequests:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "请求过于频繁，请稍后再试"
              retryAfter:
                type: integer
                description: 建议重试间隔（秒）
...